fix: dx check now respects files to ignore (e.g. .gitignore)
closes #3751
